chore: rename to depth service
